lambda_functions/silver_transformation/requirements.txt
boto3==1.34.0
python-dateutil==2.8.2
